+++ src/toisto/model/language/concept_factory.py
@@ -14,12 +14,12 @@ from .grammar import GrammaticalCategory
 from .label import label_factory
 
 CommonReferenceLevelDict = dict[CommonReferenceLevel, CommonReferenceLevelSource | list[CommonReferenceLevelSource]]
-UsesListOrString = ConceptId | list[ConceptId]
-UsesDictOrListOrString = dict[Language, UsesListOrString] | UsesListOrString
-MetaData = Literal["level", "uses"]
+ConceptIdListOrString = ConceptId | list[ConceptId]
+ConceptIdDictOrListOrString = dict[Language, ConceptIdListOrString] | ConceptIdListOrString
+MetaData = Literal["level", "antonym", "uses"]
 LeafConceptDict = dict[
     Language | MetaData,
-    ConceptId | list[ConceptId] | UsesDictOrListOrString | CommonReferenceLevelDict,
+    ConceptId | list[ConceptId] | ConceptIdDictOrListOrString | CommonReferenceLevelDict,
 ]
 CompositeConceptDict = dict[
     GrammaticalCategory | MetaData,
@@ -43,14 +43,23 @@ class ConceptFactory:
         """Create a composite concept from a composite concept dict."""
         constituent_concepts = []
         uses = cast(CompositeConceptDict, dict(uses=self.get_uses()))
+        antonyms = self.get_antonym_concepts()
         levels = self.get_levels()
         for category in self.get_grammatical_categories():
             constituent_concept_id = ConceptId(f"{self.concept_id}/{category}")
-            constituent_concept_dict = cast(CompositeConceptDict, self.concept_dict)[category] | uses
+            antonym_concept_ids = [ConceptId(f"{antonym}/{category}") for antonym in antonyms]
+            antonyms_dict = dict(antonym=antonym_concept_ids)
+            constituent_concept_dict = cast(CompositeConceptDict, self.concept_dict)[category] | uses | antonyms_dict
             constituent_concept_dict.setdefault("level", levels)
             concept_factory = self.__class__(constituent_concept_id, cast(ConceptDict, constituent_concept_dict))
             constituent_concepts.append(concept_factory.create_concept())
-        return Concept(self.concept_id, self.get_used_concepts(), tuple(constituent_concepts), level=self.get_level())
+        return Concept(
+            self.concept_id,
+            tuple(constituent_concepts),
+            self.get_used_concepts(),
+            self.get_antonym_concepts(),
+            level=self.get_level(),
+        )
 
     def leaf_concept(self) -> Concept:
         """Create a leaf concept from a leaf concept dict."""
@@ -59,7 +68,14 @@ class ConceptFactory:
             for key, value in self.concept_dict.items()
             if key in get_args(Language)
         }
-        return Concept(self.concept_id, self.get_used_concepts(), (), labels, self.get_level())
+        return Concept(
+            self.concept_id,
+            (),
+            self.get_used_concepts(),
+            self.get_antonym_concepts(),
+            labels,
+            self.get_level(),
+        )
 
     def get_grammatical_categories(self) -> tuple[GrammaticalCategory, ...]:
         """Retrieve the grammatical categories from the concept dict."""
@@ -77,9 +93,18 @@ class ConceptFactory:
             for language, concept_ids in uses.items()
         }
 
-    def get_uses(self) -> UsesDictOrListOrString:
+    def get_uses(self) -> ConceptIdDictOrListOrString:
         """Get the uses from the concept dict."""
-        return cast(UsesDictOrListOrString, self.concept_dict.get("uses", {}))
+        return cast(ConceptIdDictOrListOrString, self.concept_dict.get("uses", {}))
+
+    def get_antonym_concepts(self) -> tuple[ConceptId, ...]:
+        """Return the antonym concepts."""
+        antonyms = self.get_antonyms()
+        return tuple(antonyms) if isinstance(antonyms, list) else (antonyms,)
+
+    def get_antonyms(self) -> ConceptIdListOrString:
+        """Get the antonyms from the concept dict."""
+        return cast(ConceptIdListOrString, self.concept_dict.get("antonym", []))
 
     def get_levels(self) -> CommonReferenceLevelDict:
         """Get the Common Reference Levels from the concept dict."""