+++ src/toisto/model/quiz/quiz_factory.py
@@ -37,7 +37,11 @@ class QuizFactory:
         """Create the quizzes for a leaf concept."""
         translation_quizzes = self.translation_quizzes(concept, previous_quizzes)
         listening_quizzes = self.listening_quizzes(concept, translation_quizzes.copy() | previous_quizzes)
-        return translation_quizzes | listening_quizzes
+        antonym_quizzes = self.antonym_quizzes(
+            concept,
+            translation_quizzes.copy() | listening_quizzes.copy() | previous_quizzes,
+        )
+        return translation_quizzes | listening_quizzes | antonym_quizzes
 
     def translation_quizzes(self, concept: Concept, previous_quizzes: Quizzes) -> Quizzes:
         """Create translation quizzes for the concept."""
@@ -80,6 +84,30 @@ class QuizFactory:
                 Quiz(concept, target_language, target_language, label1, (label2,), quiz_types, blocked_by, meanings)
                 for label1, label2 in zip(labels1, labels2, strict=False)
                 if label1 != label2
+            )
+        return quizzes
+
+    def antonym_quizzes(self, concept: Concept, previous_quizzes: Quizzes) -> Quizzes:
+        """Create antonym quizzes for the concept."""
+        target_language, source_language = self.target_language, self.source_language
+        labels = concept.labels(target_language)
+        blocked_by = tuple(previous_quizzes)
+        quizzes = Quizzes()
+        for antonym in concept.antonym_concepts:
+            antonym_labels = antonym.labels(target_language)
+            meanings = concept.meanings(source_language) + antonym.meanings(source_language)
+            quizzes |= Quizzes(
+                Quiz(
+                    concept,
+                    target_language,
+                    target_language,
+                    label,
+                    antonym_labels,
+                    ("antonym",),
+                    blocked_by,
+                    meanings,
+                )
+                for label in labels
             )
         return quizzes