+++ tests/toisto/model/language/test_concept_factory.py
@@ -30,6 +30,42 @@ class ConcepFactoryTest(ToistoTestCase):
         self.assertEqual(("year",), concept.used_concepts("fi"))
         self.assertEqual((), concept.used_concepts("en"))
 
+    def test_antonym(self):
+        """Test that a concept can have an antonym concept."""
+        big = self.create_concept("big", dict(antonym="small", en="Big"))
+        small = self.create_concept("small", dict(en="Small"))
+        self.assertEqual((small,), big.antonym_concepts)
+
+    def test_multiple_antonyms(self):
+        """Test that a concept can have multiple antonyms."""
+        big = self.create_concept("big", dict(antonym=["small", "little"], en="Big"))
+        little = self.create_concept("little", dict(en="Little"))
+        small = self.create_concept("small", dict(en="Small"))
+        self.assertEqual((small, little), big.antonym_concepts)
+
+    def test_antonyms_of_composite(self):
+        """Test that a composite concept can have an antonym."""
+        big = self.create_concept(
+            "big",
+            {
+                "antonym": "small",
+                "positive degree": dict(en="Big"),
+                "comparative degree": dict(en="Bigger"),
+                "superlative degree": dict(en="Biggest"),
+            },
+        )
+        small = self.create_concept(
+            "small",
+            {
+                "positive degree": dict(en="Small"),
+                "comparative degree": dict(en="Smaller"),
+                "superlative degree": dict(en="Smallest"),
+            },
+        )
+        self.assertEqual((small,), big.antonym_concepts)
+        for index in range(3):
+            self.assertEqual((small.constituent_concepts[index],), big.constituent_concepts[index].antonym_concepts)
+
     def test_level(self):
         """Test that a concept can have a level."""
         concept = self.create_concept("one", dict(level=dict(A1="EP"), fi="Yksi", nl="Eén"))